vdk-jobs-troubleshooting: Add thread-dump utility
This change adds the implementation of a thread-dump utility to the
vdk-jobs-troubleshooting plugin.

The utility uses an http server, through which an administrator is able
to force a stacktrace dump of all threads used by the python process of
the data job. The server is bound to a port on the localhost, so to get
the stacktrace, one needs to be attached to the data job pod.

Testing Done: Added unit tests for the utility registry, and tested the
plugin itself locally by running a simple data job and examining the
execution logs.

Signed-off-by: Andon Andonov <andonova@vmware.com>