vdk: update changelog instructions
In time it has proven that maintaining a separate Changelog file is not efficient way of documenting changes. It's very rarely updated. And often the quality is not particuarly high as most users I have shown it to do not understand it.
We also have a lot of components and it's hard to track changes across all components distinct changelog files. Unoficially we had not really be using/udpating changelogs for sometime. 

This PR is abolishing this practice officially by removing it from contributing guide. 

After that I will review changelog files if there's anyh point of keeping some for information purposes and remove ones that are unnecessary.