vdk-test-utils: add cli_assert_output_contains (#1540)
* vdk-test-utils: add cli_assert_output_contains

We may need to verify that something exists in the output or some other
verification. In the same time, we'd like to make use of cli_assert for more
detailed output in case of an error (prints output logs, thread dump of
exception).

This is extending `cli_assert` method to support passing custom message
And adding new method cli_assert_output_contains which would test if
the output string contains a substring. As this is a frequently used
verification method in a lot of plugins.

new unit tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>