[vdk-plugins] vdk-meta-jobs: Fix TimeoutErrors in start_job (#1527)
Currently, if the network is unstable while executions of data jobs are started
by a meta-job, timeout exceptions may occur, which are not properly handled.

This change improves the robustness of the vdk-meta-jobs plugin by introducing logic
to handle urllib3.exceptions.TimeoutError cases, which may sporadically occur when
the network or the Control Service is unstable.

The new logic adds a retry mechanism to `start_job()` in the TrackingDataJobExecutor
class through new method called `start_new_job_execution()`. The idea is that the
network instability mey be temporary or even an one-off event, and the meta-job
should not fail. In addition, a new api call is introduced to get the list of all
recent executions and filter out everything but the latest execution. This is needed
as, even though a timeout error has been raised, that does not mean the operation
was unsuccessful. In fact, in most cases, a data job execution has been started,
and the timeout has been caused by the response from the Control Service not arriving in
time, thus starting yet another execution might actually be undesirable.

Testing done: Unit and Integration tests

Signed-off-by: Andon Andonov <andonova@vmware.com>
Co-authored-by: dakodakov <ddakov@vmware.com>