control-service: Expose v1CronJob flag as env var
Currently, if a user decides to use the `batch/v1` version of the kubernetes
cronjob API, they need to set the `datajobs.control.k8s.k8sSupportsV1CronJob`
application property to true. This, however, is a code change and applies to
all deployment environments, which means that if the user has separate test
and production setups, they cannot switch first one of the environments and
then the second, they need to switch them both at the same time.

This change introduces an environment variable `DATAJOBS_SUPPORT_V1CRONJOB`,
and adds support for setting it through the helm chart. The change, therefore,
allows for the cronjob API to be switched per environment.

Testing Done: Tested locally against a local kind cluster.

Signed-off-by: Andon Andonov <andonova@vmware.com>