versatile-data-kit: update readme with mail list
The mailing list was missing "mailto:" so it was not opening an mail but trying to redirect . 

Our mailing list posted in https://github.com/vmware/versatile-data-kit#contacts is being used for project development/maintainer purposes (Gitlab CICD failure notifications,  expiring tokens, etc) so I will update the doc to reflect this. Also joining it is 3 step process and is not that easy (you need to click, then send an email and finally someone from the team manually need to add you to teh group)

We need mailing list for users that only a click.