versatile-data-kit: update readme with mail list (#1561)
Our mailing list posted in
https://github.com/vmware/versatile-data-kit#contacts is being used for
project development/maintainer purposes (Gitlab CICD failure
notifications, expiring tokens, etc) so I will update the doc to reflect
this. Also joining it is 3 step process and is not that easy (you need
to click, then send an email and finally someone from the team manually
need to add you to teh group)
We need mailing list for users that only a click but that is separate
issue.

Here I am updating it to make clear its usage. 
The mailing list was missing "mailto:" so it was not opening mail but
trying to redirect browser and this is fixed also.

Signed-off-by: Antoni Ivnaov <aivanov@vmware.com>