vdk-jupyter: modify the way we work with notebooks in notebook plugin  (#1564)
Why:
Currently we offer this way of working with SQL and Python cells: 

![working-with_notebook-steps](https://user-images.githubusercontent.com/87015481/215505916-252d8550-280e-4503-8d59-9dcc07e8c1a0.png)

There are the reasons why we need to change it:
* Even though it looks pretty and easy to use, this parsing method is
error-prone. Since we read the whole notebook file as a JSON and get the
code as string, it becomes really hard to tests the cells which include
magic (for example when the cell includes both magic and python code).
And vdl-core does not work with iPython.
* Since the users does not use the job_input in the %sql cells this
requires them to create new connections to the database by themselves -
they cannot use the connections created by VDK
* Since these %sql cells does not include job_input, to make them run
with vdk we need to do changes on the cell execution of Jupyter, which
will take a lot of time and will be hard to be achieved. Also, changing
the cell execution cannot be done via ipython extension, we will need to
do modifications using jupyterlab-extension. Having one solution
separated into two totally different packages does not seem to be a good
solution.

What:
Changes on the way we read the JSON schema of a notebook.
<img width="1359" alt="Screenshot 2023-01-30 at 10 44 14"
src="https://user-images.githubusercontent.com/87015481/215506575-1a45df22-db3d-42f0-b0e0-b0c60f478e04.png">

As you can see, in this notebook we use job_input variable to execute a
SQL query instead of %sql magic.
* It does exactly the same thing as the solution above but it is a safer
option to chose because:
* Since the SQL code directly turn into python code which calls for SQL
execute, this solution is more error-proof (since everything for VDK
will be python code, we will not need to parse SQL code by hand, if
there is a problem with the SQL statement we will get an exception and
throw it)
* The users use the job_input variable directly, so they can use the VDK
connections directly, too
All the solution remains in one place instead of being separated into
different packages


Testing: Unit tests

Signed-off-by: Duygu Hasan [hduygu@vmware.com](mailto:hduygu@vmware.com)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>