vdk-core: track configuration sensitivity
https://github.com/vmware/versatile-data-kit/issues/436

Why?
Enables vdk-core and plugins to mark configurations as sensitive.
Append 'This option is marked as sensitive' to descriptions of
configurations that are marked as sensitive.
Examples of sensitive configurations are passwords, api keys, etc.

How has this been tested?
Add a sensitive option and check for it in the existing config-help unit test

What type of change are you making?
New feature (non-breaking change which adds functionality)

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>