control-service: Cronjob API backwards compatibility
Currently there is no backwards compatibility support when a user
switches to the V1 Kubernetes Cronjob API (from the currently default
V1beta1). This results in situations, where if there are deployed data
jobs which use th v1beta1 API, and a switch is made to the V1 API, these
jobs are suddenly shown as `NOT DEPLOYED`, and cannot be properly managed.

This change adds backwards compatibility support in the Kubernetes Service,
to allow for a switch to V1 Cronjob API in clusters, where V1beta1 API cronjobs
are deployed.

Testing Done: Unit and Integration tests (new and existing).

Signed-off-by: Andon Andonov <andonova@vmware.com>