vdk-meta-jobs: Better error message for misspelled job name
If a user misspells a job name in the depends_on section,
currently they will receive an error informing them that
the job is not registered. Job registration is an
implementation detail and is not a concept users would
be familiar with, so this PR amends the error message
to tell the user that they might have mistyped the job name.

Signed-off-by: Gabriel Georgiev <gageorgiev@vmware.com>