versatile-data-kit: simplify release process
Having to enumerate every single package is time-consuming and is never used by anyone. 
Those versions are rather arbitrary and point-to-time snapshots at the time of release announcement which could be found if needed from the GItlab CI and does not represent anything special (like a more stable or battle-tested version). 

So instead we can simplify the process and only include install instructions.