vdk-trino: stabilize vdk-trino tests (#1677)
This change is fixing vdk-trino tests which are failing regularly in the
nightly builds . See
https://github.com/vmware/versatile-data-kit/issues/1559

The tests were failing intermittently for two main reasons 
* Docker failures which are being fixed by adopting testcontainers (in
favour of pytest-docker)
* Hidden dependencies between test cases (because different test use
same schema and tables on shared db instance)

Until now we used pytest-docker to start service container, this is
switching it to testcontainers-python.

There are multiple advantages of using it.

* Supposedly Testcontainers is built with performance in mind and uses a
variety of techniques to minimize container startup time and resource
usage like image reuse, container reuse, asynchronous operations.
* It is far more popular lately -
https://hugovk.github.io/top-pypi-packages/ is a site I frequently use
to decide which library is likely to be more stable/used . It is showing
testcontainers is 2000 places ahead of pytest-docker
* Testcontainers provides a wider range of features for managing
containers, such as starting and stopping them, waiting for them to
become available, and exposing their ports to the host system
* Testcontainers-Python allows you to define and configure containers
using Python code instead of yaml which pytest-docker required.

Testing Done: ran the test suite of vdk-trino a few times and it was
more stable and quick. I have not run actual benchmarks and recorded
times but it seems about twice faster possibly. We will verify this as
part of the nightly builds.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>