frontend: Rename vdk/shared to versatiledatakit/shared (#1694)
### Why
    
Fixes the broken CI release job for vdk/shared. The account we publish
to in
the registry is called versatiledatakit. Registry only supports scoped
packages,
e.g. @something/something-else. We can't publish a package to that
account,
unless it matches the account, so it should be called
@versatiledatakit/shared
and not @vdk/shared. Otherwise, publishing fails
    
### What
    
Rename all references for vdk/shared to versatiledatakit/shared
Add package-lock.json to .gitignore
Change description of package to not mention Taurus
    
### How has this been tested
    
Tested the scripts by running them locally
    
### What type of change are you making
    
Bugfix

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>