frontend: data-pipelines build and test running CI/CD (#1671)
We need to build and run unit tests of data-pipelines library and ui
application that uses it. The component is dependent on
shared-components library.

Added Bash scripts to
(1) link if any shared-components rebuild was made and is about to be
published, or otherwise default to latest already published one
(2) build the data-pipelines library and link @vdk/data-pipelines 
(3) build the UI application
(4) lint the subprojects to verify format
(5) run the unit tests
One bash script is meant to be used to satisfy build prerequisites, the
other one to rebuild and relink continuously, needed during development.
The proposed implementation triggers a data-pipelines rebuild (+test),
upon any shared-components change, as well as data-pipelines changes.
A README.md link fix to the VEP.

Testing done: run the install script (that evaluates the build script)
on a clean setup, and verified the unit tests were successful. Also, did
start and browse the application manually, to check its operational.

---------

Signed-off-by: ivakoleva <iva.koleva@clearcode.bg>