vdk-quickstart: remove use of global varaibles in CI
The use of global variables in GitLab CI can create problems when
multiple CI files are being imported, as variables are not scoped to a
specific file. This can lead to unexpected variables being set, causing
issues in the CI/CD pipeline.

This is what is happening wiht quickstart-vdk test -
https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/3901914540

It expected that there's python installed because it's supposed to use
python image but some other file overrides this image to docker and
causes the test to fail.

This changes quickstart-vdk to use job-level variables.

In future changes I will be changing all CIs to use  job-level variables
or environment variables to avoid conflicts and ensure proper scoping.
Or ensure that global variabels and labels are prefixed with the project
or plugin name.

Testing Done: this PR CI

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>