vdk-plugins: generate test report
The Gitlab CI was not showing the run plugin tests because it was not
finding the test report file.

This is fixing that by specify full path.

Testing Done: this Gitlab CI pipeline would be able to see the report in
the UI . See https://tinyurl.com/283p9mjk
While before not: https://tinyurl.com/yeyufrfx

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>