frontend: data-pipelines sync to freezepoint
Had to fetch the latest changes from the internal repo contributed,
to unify the diverged with the public repo.

Synced the functional changes of frontend/data-pipelines/gui/projects/*
The frontend/data-pipelines/gui/e2e/* sync is scoped with PR
https://github.com/vmware/versatile-data-kit/pull/1705

Testing done: did run locally the install script; the majority of errors
seem to be due versatiledatakit/shared sync prerequisite needed.
When the PR with those is ready, I'll rebase my feature branch onto it,
to verify.

Signed-off-by: ivakoleva <iva.koleva@clearcode.bg>