Control Service: Add supported python version configuration
As part of the effort to support different python versions for data job
deployments in Versatile Data Kit, we need to be able to configure what
python versions are supported by the Control Service.

This change adds new configuration to the Control Service's helm chart,
which allows administrators of the service to specify what python versions
are allowed for deployments and what vdk and job base images are to be used
for the corresponding python versions. The configuration is in the form of
a map as described in VEP-1739 and is injected in the Control Service
deployment as a `DATAJOBS_DEPLOYMENT_SUPPORTED_PYTHON_VERSIONS`
environment variable.

The configuration is also added to the service's application properties
files.

Additionally, a set of methods is implemented to support a variety of use
cases related to accecssing and managing the supported python versions
configuration.

Testing Done: Unit tests added.

Signed-off-by: Andon Andonov <andonova@vmware.com>