vdk-plugins: local differential privacy plugin  (#1760)
# Why?
In the last few years we have seen the emergence of Big Data within the
data space. From making and validating business decisions to training
machine learning models, data has shown itself to be vital to the
operation of many businesses.

However, this trend has coincided with a greater interest in data
privacy. Users are interested more than ever in protecting their
personal data.

This raises the question, how can we collect as much data as possible,
while respecting our users’ privacy? Different answers have been given
to this query.

Historically, data teams would simply obfuscate personally identifiable
information, such as names for example, in their datasets; however this
is not sufficient for protecting a user’s privacy, as a malicious actor
can easily pick out quasi-identifying elements within the dataset.

Differential privacy has been proposed as a solution to this problem.
Simply put, the idea is that we introduce some controlled amount of
statistical noise within the data that we collect; this means that the
data cannot be used reliably to identify users, while at the same time
preserving its usefulness for decision making or training ML models.

However, a main obstacle to differential privacy’s popularisation is the
lack of tooling to support it: while some libraries for data processing
or plugins for databases exist, there are few integrations with data
engineering infrastructure frameworks that can be picked up and easily
used.

# What?
This PR contains a plugin that can be used to add statistically noise to
enum like data and/or boolean data during the ingestion phase.

# How has this been tested?
new unit tests added

# What type of change are you making?
New Feature

# What is missing?
1. at the moment the amount of noise is not configurable but the code is
written in a way to make it easily configurable in the future.

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>