control service: Dynamically set job base image in builder
Currently, the job base image used in job builders is set in the `deploymentDataJobBaseImage`
application property. This is set once at Control Service deploy and requires re-deploy of the
service to be changed.

As part of the initiative to support multiple python versions for data job deployments, a new
application property, `supportedPythonVersions`, was introduced in
https://github.com/vmware/versatile-data-kit/commit/967898fddec3e0515f1e5817bc2211cebc1b2d92.
Additionally, a new API property, `python_version` was introduced in
https://github.com/vmware/versatile-data-kit/commit/5086fa3d2fde0bb9db38e3ca1928764c95cf3900
to allow users to set which python version to be used for job deployments.

With all this done, this change updates the JobImageBuilder to use the above-mentioned configuration
to set the job base image dynamically, without requiring Control Service redeploy.

Testing Done: New and existing tests.

Signed-off-by: Andon Andonov <andonova@vmware.com>