control-service: protect against v1 cron batch not being supported (#1777)
# Why
When we merged this PR
https://github.com/vmware/versatile-data-kit/pull/1580/files we
unintentionally broke compatibility with version 1.19 and 1.20.

We always check for v1 versions of batch jobs. checking for that version
on clusters <1.21 will result in errors.

# What
Within this PR we add a guard around that statement and swallow the
error if it happens.

# How has this been tested
installed old version of k8s locally and ran integration tests against
it.
Within the class DataJobDeploymentCrudIT I also changed the property
`atajobs.control.k8s.k8sSupportsV1CronJob` to `false`. As this is the
config that should be set for 1.20.
All tests are green. 

I also ran the integration tests on main against the old cluster to make
sure it catches the bug and it does result in many failed tests.
 

# Not handled in this PR
How do we prevent issues like this in the future? 
Maybe integration tests run against many versions of k8s?

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: github-actions <>