control-service: Add python_version to Control Service API
As part of the effort to provide users with the ability to specify the python
release version, which they want their data jobs to be deployed with, we need to
expose a property in the Control Service's API and ensure that the configuration
is stored in the database and the kubernetes objects.

This change exposes a `python_version` property in the api schemas of the
DataJobDeployment and DataJobDeploymentStatus components, and propagates the property
to the database model and kubernetes resources. All of this is done as one change,
to avoid unexpected behaviour or regressions.

Testing Done: New and modified existing tests.

Signed-off-by: Andon Andonov <andonova@vmware.com>