specs: update Multiple Python versions VEP summary (#1792)
As recommended in the [VEP
Template](https://github.com/vmware/versatile-data-kit/edit/main/specs/NNNN-template/README.md)
, the summary could be "used as user-focused documentation such as
release notes or a (customer-facing) development roadmap. The tone and
content of the `Summary` section should be useful for a wide audience."

The current summary had some unnecessary technical details and this
change aims to overall make it easier for non-technical or casual
users/readers to understand.