vdk-plugins: test only oldest and newest supported python version
As part of my effort to stabilize our Nightly builds and the CICD
cluster I am removing testing middle versions.

We test with the oldest supported version to make sure we do not use a
new feature added lately which would break compatability.

We test with newest version to make sure that it runs with newest
version. Although the python itself is very likely to be backwords
compatible (for minor release), there are sometimes small changes that
are not and same cannot be told for 3th party dependecies we may use
which may not support latest version fully.

Testing Done: this CI

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>