specs: VEP-1739 Update status and reorganise document
This change updates the status of `VEP-1739 Multiple Python Versions` to `implementable`,
as work is underway to implement the proposal. Additionally, the document is re-organised
and the options discussed for supported python versions configuration are moved to the
Alternatives section, and only the chosen solution remains in the Detailed Design part.

Testing Done: N/A

Signed-off-by: Andon Andonov <andonova@vmware.com>