vdk-notebook: add hook for saving error information into json file (#1842)
What:
Added a hook that saves error information from job_steps into error.json
file.
Also, for more informative error logs caused from cells from .ipynb
file, added the id of the failing cell to the logs. (which will be used
later from the Jupyter UI for marking the cell)

Why: The main reason for the changes is that in the Jupyter UI we want
to introduce dialogs for job run status which consist of shorter message
for errors (not all of the logs). But in general, it is a good option to
have a place where the reason for the failure is kept (the user does not
have to go over all of the logs to see which step is failing).


Signed-off-by: Duygu Hasan [hduygu@vmware.com](mailto:hduygu@vmware.com)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>