frontend: Omit optional dependencies when testing
Why?

The CICD scripts pull data-pipelines and shared packages
from registry instead of using the linked packages. This
happens because optional dependencies are not omitted on
running npm install

What?

Add the omit flag to npm install for both scripts

How was this tested?

1. Ran tests locally
2. CI/CD

What type of change are you making?

Bugfix, non-breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>