frontend: publish docker image for ui
Why?

Part of the wider effort to add the UI to quickstart-vdk

What?

Build and publish a docker image for the UI as part
of the release step in the CI/CD pipeline. Image tags
follow the same versioning schema as the data-pipelines
package in the npm registry

How has this been tested?

Ran the gitlab script locally
https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/4148986837

What type of change are you making?

New feature (non-breaking change which adds functionality)

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>