control-service: Dynamically set vdk image in JobImageDeployer
As part of the initiative to support different python versions for data job
deployments, we need to set the vdk image used for deployments dynamically.

This change updates the JobImageDeployer component to the vdk image
dynamically. Additionally, the patchDeployment method in the Deployment service
is modified to not allow vdk images and python versions be changed on patch
requests, as patch requests do not support rebuilding of the job image, which
is necessary when multiple python versions are supported.

Testing Done: Current and new tests and integration tests.

Signed-off-by: Andon Andonov <andonova@vmware.com>