frontend: Implement changed history link for Job
Implement feature for Data Job changed history link that is
instructed with feature flag whether to show or hide, and
on which page to show.
Confirmation title and message are provided as dependency
injection to the library config.
Tested with generating local build, linking built artefacts
back to @taurus/shared and then in other libraries
and manual validated that everything works correctly.
Executed prettier for formatting new files and also
successfully executed ESLint and there are no error issues.

Signed-off-by: gorankokin <gorankokin@gmail.com>