Simplify and clarify VDK purpose in the README (#1837)
This change aims to provide a better, simpler, and easier-to-read way to
understand what VDK does.

The image was one of the proposed ways to show it. I can also do the
same with text and some complex formatting. Open for feedback.

The way it's intended to be more clear - short sentences
More readable - white space, formatting
Simple - the division between two components.

---------

Co-authored-by: Antoni Ivanov <aivanov@vmware.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>