control-service: Expose supported python versions in helm
As part of the initiative to support multiple python versions in data job deployments,
new configurations were introduced in the form of `datajobs.deployment.supportedPythonVersions`
and `datajobs.deployment.defaultPythonVersion` configuration properties.

To facilitate proper production setup of the Control Service with the new functionality, the
configuration needs to be exposed in the helm chart.

This change updates the values.yaml and deployment.yaml files of the Control Service helm chart
to expose the supportedPythonVersions and defaultPythonVersion configurations and set them as
environment variables.

Testing Done: CI/CD and local run of `helm lint` on the chart.

Signed-off-by: Andon Andonov <andonova@vmware.com>