frontend: visibility of app components is configurable
Why?

As part of releasing the UI to the community,
we should be able to configure the visibility of
certain components that are not intended for
external use

What?

Component visibility for explore page is configurable
through AppConfigService
Routes are configurable through AppConfigService

How was this tested?

Locally plus component specs

What type of change is this?

Feature, breaking

Signed-off-by: Dilyan Marinov <mdilyan@vmware.com>