control-service: new python client. (#1983)
# Why 
The newest python api client is broken and the last working one is very
old.

# What
In this PR I use the python-nextgen generator instead of the python
generator. The behaviour is much more correct and predictable.

There are two bugs in the generated code. 
For both these I have created tickets in the openapi generator github
project.

I have added string replace workarounds in the gradle step to fix these
bugs before we publish the pip package.


next steps(will be tackled in the next few weeks not days):
https://github.com/vmware/versatile-data-kit/issues/1982

closes: https://github.com/vmware/versatile-data-kit/issues/1979


# How has this been tested? 
Locally. 
There is actually a draft branch where I import this new version into
the cli project and run tests against it to confirm everything is green:
https://github.com/vmware/versatile-data-kit/pull/1969

This will always fail in github because 1.0.9 is not released yet till
this PR is merged. Then I will open a new PR where I import into cli
tool.

Signed-off-by: murphp15 <murphp15@tcd.ie>