control-service: Rework supported python version logic
Currently, there are two groups of configurations to set the vdk and job base images
used in data job deployments. The old way, through the `datajobs.deployment.dataJobBaseImage`
and `datajobs.vdk.image`, and the new way, through the support for multiple python versions in
data job deployments.

The two groups of configurations are in conflict, as the old configuration options take precedence
and cleanly migrating to the support for multiple python versions is practically impossible.

This change removes the default values for the `datajobs.deployment.supportedPythonVersions` and
`datajobs.deployment.defaultPythonVersion` properties, and udates the service logic so that when these
properties are set in an environment, they take precedence over the old configuration.

Testing Done: CI/CD

Signed-off-by: Andon Andonov <andonova@vmware.com>