control-service: custom app config values can we passed to helm. (#2004)
# Why
We want to be able to change the frontend config depending on the
situation it is deployed in.
This lets users provide their own appConfig file to the frontend. 
An example of this is toggling auth on and off. 

# What
Use k8s volumemounts to load in the config file.

# How has this been tested?
Deployed the helm locally.

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>