control-service: logs url can include team name (#2013)
# Why
Often the URL where the logs are stored will need the team name. 
An example of this is when using our own api to serve the logs.

# What
team name can now be used in the URL 

# How has this been tested 
Updated unit tests. 
In the next two days I will be incorporating this change into quickstart
vdk.

I can wait till it is incorporated in that to merge this if ppl don't
feel comfortable with the unit test.

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: github-actions <>