control-service: remove unused parameters (#2016)
# Why 
I am trying to improve the stability of the integration tests. I have
seen a way of improving them but it requires some code changes.

Before making those code changes I want to push this clean up. as it is
very hard to build on the functions that are currently there.


I removed 4 parameters that are never used. 
I also removed podLabels which is used but it is always an empty map. So
it never did anything.
 
# How was this tested
Unit and integration tests.

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: github-actions <>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>