vdk-plugin-control-cli: Add python version to log
This change updates the log message at the start of a new execution by adding
the python version of the execution environment. This information can be used for
troubleshooting purposes, when a job is deployed and executed with different
python distributions, e.g., if it is first deployed and executed with python 3.7,
but then re-deployed and executed with python 3.9, and there are issues with one
of the executions but not with the other.

Testing Done: Executed a data job locally with `vdk run ...`, and verified that
the python version is properly logged.

Signed-off-by: Andon Andonov <andonova@vmware.com>