control-service: fix oom tests (#2028)
# Why
In this PR I fix a very unstable oom test. 
The test is very unstable for a number of reasons. 

1. sometime the job actually succeeds, it seems to be able to compelte
with in the memory limits after a number of retries.
We can put the memory requirement any lower to ensure an error because
6mi is the lower limit supported by k8s.
2. on of the assertions we were making is that the logs are there.
However often if the jobs is killed at the very start there are no logs.


# What 
I fix it by using a cronjob template which doesn't let the job restart.
This means if it fails once it will report as failed.


# How has this been tested?
integration tests.

---------

Signed-off-by: murphp15 <murphp15@tcd.ie>
Co-authored-by: github-actions <>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>