vdk-notebook: make it main distribution of all jupyter plugins
Currently in order to get the full notebook integration one need to
install 3 plugins - vdk-notebook, vdk-ipython and vdk-jypyter. This kind
of separation is very good both for development purposes and to allow
users to choose what they need.

For first time users though we'd like them to have access to everything
and by default to be able to use all functionality. In that case it's
better to have a single distribution that install all.

I played around with introducing aggregation type of plugin (like
quickstart-vdk) but ultimately seems like too many plugins and we can
re-use one of hte existing ones. I choose vdk-notebook because hte name
is the most generic one and can be easily described to include jupyter
and ipython integraiton.

The change improves the README accordingly.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>